+++ sample/src/main/kotlin/com/sonai/ssiv/test/imagedisplay/ImageDisplayRotateFragment.kt
@@ -12,21 +12,19 @@ import com.sonai.ssiv.test.R
 class ImageDisplayRotateFragment : Fragment() {
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View? {
         val rootView = inflater.inflate(R.layout.imagedisplay_rotate_fragment, container, false)
         val imageView = rootView.findViewById<SubsamplingScaleImageView>(R.id.imageView)
         imageView.setImage(ImageSource.asset("swissroad.jpg"))
-        imageView.setOrientation(90)
+        imageView.orientation = 90
         val activity = activity as? ImageDisplayActivity
         if (activity != null) {
             rootView.findViewById<View>(R.id.previous).setOnClickListener { activity.previous() }
             rootView.findViewById<View>(R.id.next).setOnClickListener { activity.next() }
         }
         rootView.findViewById<View>(R.id.rotate).setOnClickListener {
-            imageView.setOrientation((imageView.getOrientation() + 90) % 360)
+            imageView.orientation = (imageView.orientation + 90) % 360
         }
         return rootView
     }